+++ HISTORY.md
@@ -3,6 +3,7 @@ Release Date: 2020-TBA
 
 * Data parity with faker.js. Deterministic sequences using `Internet.Avatar()` may have changed.
 * Add support for .snupkg NuGet Symbol Server Packages via SourceLink.
+* Added Randomizer.Enums() that makes selecting a subset of enum values easier.
 
 ## v28.4.4
 Release Date: 2019-12-10